LayoutTests/fast/css/image-set-setting-expected.txt

Test the setting of the -webkit-image-set function.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Single value for background-image : url('http://www.webkit.org/a') 1x
PASS testComputedStyle("background-image", "url('http://www.webkit.org/a') 1x") is "-webkit-image-set(url(http://www.webkit.org/a), 1)"

Multiple values for background-image : url('http://www.webkit.org/a') 1x, url('http://www.webkit.org/b') 2x
PASS testComputedStyle("background-image", "url('http://www.webkit.org/a') 1x, url('http://www.webkit.org/b') 2x") is "-webkit-image-set(url(http://www.webkit.org/a), 1, url(http://www.webkit.org/b), 2)"

Multiple values for background-image, out of order : url('http://www.webkit.org/c') 3x, url('http://www.webkit.org/b') 2x, url('http://www.webkit.org/a') 1x
PASS testComputedStyle("background-image", "url('http://www.webkit.org/c') 3x, url('http://www.webkit.org/b') 2x, url('http://www.webkit.org/a') 1x") is "-webkit-image-set(url(http://www.webkit.org/c), 3, url(http://www.webkit.org/b), 2, url(http://www.webkit.org/a), 1)"
PASS successfullyParsed is true

TEST COMPLETE
